+++ Wrappers/Python/cil/framework/framework.py
@@ -2771,9 +2771,9 @@ class DataContainer(object):
         return self.array
 
 
-    def get_slice(self,**kw):
+    def get_slice(self, **kw):
         '''
-        Returns a new DataContainer containing a single slice of in the requested direction. \
+        Returns a new DataContainer containing a single slice in the requested direction. \
         Pass keyword arguments <dimension label>=index
         '''
         new_array = None